refactor: organizes Badge exports in their respective folders
