+++ src/components/ui/badge/Badge.vue
@@ -1,14 +1,6 @@
 <script setup lang="ts">
 import { computed } from 'vue';
-import type { Size } from '../../../types';
-
-export interface BadgeProps {
-  label?: string | number;
-  pill?: boolean;
-  size?: Size;
-  noWrap?: boolean;
-  variant?: 'highlight' | 'warning' | 'success' | 'critical' | 'default';
-}
+import type { BadgeProps } from './types';
 
 const props = withDefaults(defineProps<BadgeProps>(), {
   size: 'md',
