+++ src/components/ui/badge/index.ts
@@ -1,0 +1,2 @@
+export { default as Badge } from './Badge.vue';
+export type { BadgeProps } from './types';
